+++ homeworks/hw1/prog/hw1_str.c
@@ -11,13 +11,13 @@ int str_manip(char *str, char *substr) {
     printf("str %s\n", str);
     char low[str_len + 1];
     strncpy(low, str, str_len + 1);
-    for (int i = 0; i < str_len; i++) {
+    for (size_t i = 0; i < str_len; i++) {
         low[i] = tolower(low[i]);
     }
     char rev[str_len * 2 + 1];
     rev[0] = '\0';
     strcat(rev, low);
-    for (int i = 1; i <= str_len; i++) {
+    for (size_t i = 1; i <= str_len; i++) {
         rev[str_len - 1 + i] = low[str_len - i];
     }
     rev[str_len * 2] = '\0';
@@ -25,7 +25,7 @@ int str_manip(char *str, char *substr) {
     printf("substr %s\n", substr);
     char substr_low[substr_len + 1];
     strncpy(substr_low, substr, substr_len + 1);
-    for (int i = 0; i < substr_len; i++) {
+    for (size_t i = 0; i < substr_len; i++) {
         substr_low[i] = tolower(substr_low[i]);
     }
     int ocr = 0;
